EW-173: HomePage model
